app8ThreatsAndIncidents/microservice3/requirements.txt
uvicorn
fastapi
pydantic
asyncio
aiohttp
aiosqlite
pycryptodome